MWPW-165482 milo autoblock feature
