+++ studio/src/constants.js
@@ -37,12 +37,3 @@ export const EVENT_KEYUP = 'keyup';
 export const EVENT_FRAGMENT_CHANGE = 'fragment:change';
 
 export const EVENT_OST_SELECT = 'ost-select';
-
-//TODO make that feed (excepts ALL maybe) dynamically served from milo
-export const VARIANTS = [
-    { label: 'All', value: 'all', surface: 'all' },
-    { label: 'Catalog', value: 'catalog', surface: 'acom' },
-    { label: 'Slice', value: 'ccd-slice', surface: 'ccd' },
-    { label: 'Special offers', value: 'special-offers', surface: 'acom' },
-    { label: 'Suggested', value: 'ccd-suggested', surface: 'ccd' },
-];
