+++ studio/src/editors/variant-picker.js
@@ -1,5 +1,13 @@
 import { html, LitElement } from 'lit';
-import { VARIANTS } from '../constants.js';
+
+//TODO make that feed (excepts ALL maybe) dynamically served from milo
+export const VARIANTS = [
+    { label: 'All', value: 'all', surface: 'all' },
+    { label: 'Catalog', value: 'catalog', surface: 'acom' },
+    { label: 'Slice', value: 'ccd-slice', surface: 'ccd' },
+    { label: 'Special offers', value: 'special-offers', surface: 'acom' },
+    { label: 'Suggested', value: 'ccd-suggested', surface: 'ccd' },
+];
 
 class VariantPicker extends LitElement {
     static properties = {
